% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mobileanalytics_package.R
\docType{package}
\name{paws.mobileanalytics-package}
\alias{paws.mobileanalytics}
\alias{paws.mobileanalytics-package}
\title{paws.mobileanalytics: Amazon Mobile Analytics}
\description{
Amazon Mobile Analytics is a service for
    collecting, visualizing, and understanding app usage data at scale.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
