% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mobileanalytics_operations.R
\name{put_events}
\alias{put_events}
\title{The PutEvents operation records one or more events}
\usage{
put_events(events, clientContext, clientContextEncoding = NULL)
}
\arguments{
\item{events}{An array of Event JSON objects}

\item{clientContext}{The client context including the client ID, app title, app version and package name.}

\item{clientContextEncoding}{The encoding used for the client context.}
}
\description{
The PutEvents operation records one or more events. You can have up to 1,500 unique custom events per app, any combination of up to 40 attributes and metrics per custom event, and any number of attribute or metric values.
}
\examples{

}
